src/requirements.txt
requests==2.32.3
schedule
aiohttp
PyYAML
sqlalchemy
pandas
polars
flatten_json



